requirement.txt
adb-shell[async]
opencv-python
psd-tools


notebook
numpy
matplotlib
pillow
polars
torch
torchinfo
torchvision
tqdm